// lib/types/supabase.ts
export interface definitions {
  Secret: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    Id: string;
    KeyId: string;
    Secret: string;
    Salt: string;
    AesIv: string;
    CreatedAt: string;
    ExpiresAt: string;
    SearchKey?: string;
  };
}